Tweak the intro of the package.
A quick read and the caveats confused me a little bit. It must be clear right
upfront that you can do everything with this package that you could do
without it. The caveats are not limiting the power of the API but merely the
ability to fit some corner cases in the simpler framework.